Connect to Microsoft Planner via Microsoft 365

These functions create a connection to Microsoft Planner via Microsoft 365 and saves the connection to the certeprojects package environment. The planner_*() functions allow to work with this connection.

planner_connect(
  team_name = read_secret("team.name"),
  plan_name = read_secret("planner.name"),
  ...
)

planner_bucket_create(bucket_name, account = planner_connect())

planner_buckets_list(account = planner_connect(), plain = FALSE)

planner_tasks_list(account = planner_connect(), plain = FALSE)

planner_categories_list(account = planner_connect())

planner_task_create(
  title,
  descr = NULL,
  duedate = NULL,
  requested_by = NULL,
  priority = read_secret("planner.default.priority"),
  checklist_items = NULL,
  categories = NULL,
  assigned = NULL,
  bucket_name = read_secret("planner.default.bucket"),
  account = planner_connect()
)

planner_task_update(
  old_title,
  title = NULL,
  bucket_name = NULL,
  descr = NULL,
  duedate = NULL,
  assigned = NULL,
  assigned_keep = FALSE,
  categories = NULL,
  categories_keep = FALSE,
  checklist_items = NULL,
  priority = NULL,
  account = planner_connect()
)

planner_task_id(old_title, account = planner_connect())

planner_task_request_validation(
  title,
  category_text = read_secret("planner.label.authorise"),
  account = planner_connect()
)

planner_task_validate(
  title,
  category_text = read_secret("planner.label.authorised"),
  account = planner_connect()
)

planner_user_property(
  user,
  team_name = read_secret("team.name"),
  account = planner_connect(),
  property = "id",
  as_list = FALSE
)

Arguments

team_name

name of the team

plan_name

name of the team's plan

...

arguments passed on to get_microsoft365_token()

bucket_name

name of the bucket

account

a Microsoft 365 account to use for looking up properties. This has to be an object as returned by planner_connect() or via AzureGraph::create_graph_login().

plain

return as plain names, not as Azure objects

title

title of the task

descr

a description for the task. A vector will be add as one text separated by white lines.

duedate

a date to use a due date, use FALSE to remove it

requested_by

name of the person(s) who requested the task

priority

a priority to set. Can be ranged between 0 (highest) and 10 (lowest), or: "urgent" or "dringend" for 1, "important" or "belangrijk" for 3, "medium" or "gemiddeld" or FALSE for 5, "low" or "laag" for 9. Priorities cannot be removed - the default setting is 5.

checklist_items

character vector of checklist items

categories

names of categories to add

assigned

names of members within the plan. Use FALSE to remove all assigned members.

old_title

title of the task to update

assigned_keep

add members that are set in assigned instead of replacing them, defaults to FALSE

categories_keep

add categories that are set in categories instead of replacing them, defaults to FALSE

category_text

text of the category to use

user

a user name, mail adress, or Certe login name

property

property to return, can be "id", "name" or "mail"

as_list

return the full list of members as list, split into Eigenaars (Owners) / Leden (Members). This ignores user.