Project Properties

Retrieve project properties, such as the title, folder location and project number.

project_get_current_id(ask = NULL, account = connect_planner())

project_identifier(project_number = project_get_current_id())

project_get_folder(
  project_number = project_get_current_id(),
  account = connect_planner()
)

project_get_folder_full(
  project_number = project_get_current_id(),
  projects_path = read_secret("projects.path"),
  account = connect_planner()
)

project_get_title(
  project_number = project_get_current_id(),
  account = connect_planner()
)

project_get_file(
  filename = ".*",
  project_number = project_get_current_id(),
  fixed = FALSE,
  account = connect_planner()
)

project_set_file(
  filename,
  project_number = project_get_current_id(),
  account = connect_planner()
)

project_set_folder(
  foldername,
  project_number = project_get_current_id(),
  account = connect_planner()
)

project_open_analysis_file(
  project_number = project_get_current_id(ask = TRUE),
  account = connect_planner()
)

project_open_folder(
  project_number = project_get_current_id(ask = TRUE),
  account = connect_planner()
)

project_add_qmd_skeleton(
  filename = NULL,
  project_number = project_get_current_id(),
  account = connect_planner()
)

Arguments

ask

logical to indicate whether the project number should always be asked. The default, NULL, will show a popup in interactive R sessions, allowing to search for projects. In non-interactive sessions, such as in Quarto and R Markdown, it will use the current working directory to determine the project number.

account

a Microsoft 365 account to use for looking up properties. This has to be an object as returned by connect_planner() or via AzureGraph::create_graph_login()$get_group(name)$get_plan(plan_title).

project_number

Planner project number

projects_path

location of the folder that contains all department projects

filename

name for the new Quarto file

fixed

logical to turn off regular expressions

foldername

foldername to set

Details

project_get_current_id() uses planner_task_search() to find a specific project based on any search string.

project_identifier() generates the project identifier for print on reports and in mails: a combination of the currently logged in user (in your case: 'runner '), the current date/time (format: YYMMDDHHMM), and the project number. If the project number is not available, it will only return the current user and date/time (format: YYMMDDHHMM).

project_set_folder() will create the folder if it does not exist.

project_add_qmd_skeleton() initializes a new Quarto skeleton for a project.

Examples

project_identifier(123)
#> [1] "runner-2409020859-123"