Knit R Markdown or Quarto Document using Quarto

Runs input files through Quarto, and allows Quarto to work with full paths (where Quarto itself requires relative paths). This function essentially replaces both rmarkdown::render() and quarto::quarto_render().

render(input_file, output_file = NULL, quiet = TRUE, as_job = "auto", ...)

knit(input_file, output_file = NULL, quiet = TRUE, as_job = "auto", ...)

render_sharepoint(
  input_file = NULL,
  output_file = NULL,
  project_number = project_get_current_id(),
  full_sharepoint_path = NULL,
  quiet = TRUE,
  as_job = "auto",
  account = connect_teams(),
  ...
)

Arguments

input_file

file to be rendered, can be R Markdown (.Rmd) or Quarto (.qmd), or a lot of other formats such as .md, .ipynb and many other formats that Quarto supports.

output_file

The name of the output file. If using NULL, the output filename will be based on the filename for the input file. output_file is mapped to the --output option flag of the quarto CLI. It is expected to be a filename only, not a path, relative or absolute.

quiet

Suppress warning and other messages, from R and also Quarto CLI (i.e --quiet is passed as command line).

quarto.quiet R option or R_QUARTO_QUIET environment variable can be used to globally override a function call (This can be useful to debug tool that calls quarto_* functions directly).

On Github Actions, it will always be quiet = FALSE.

as_job

Render as an RStudio background job. Default is "auto", which will render individual documents normally and projects as background jobs. Use the quarto.render_as_job R option to control the default globally.

...

arguments passed on to render()

project_number

Planner project number

full_sharepoint_path

Full path of the remote file, consisting of the project folder and the file name, such as "Project - p123/my_file.R".

account

a Microsoft 365 account to use for looking up properties. This has to be an object as returned by connect_teams() or Microsoft365R::get_team().

Value

As opposed to quarto::quarto_render(), which always returns NULL, this render() function always returns the output file name (invisibly).

Details

Functions knit() and render() are identical.

render_sharepoint() downloads a SharePoint project file to a local temporary folder, then renders it. Either input_file + project_number or full_sharepoint_path must be given.