Schedule Task (CRON-like)

This will start an expression if requirements are met, using a CRON-like syntax (https://cron.help).

schedule_task(
  minute,
  hour,
  day,
  month,
  weekday,
  user,
  expr,
  log = TRUE,
  ref_time = Sys.time()
)

Arguments

minute

one or more values between 0-59, or . or "*" for each minute

hour

one or more values between 0-23, or . or "*" for each hour

day

one or more values between 1-31, or . or "*" for each day

month

one or more values between 1-12, or . or "*" for each mpnth

weekday

one or more values between 0-7 (Sunday is both 0 and 7; Monday is 1), or . or "*" for each weekday

user

logged in user, must correspond with Sys.info()["user"]. Currently logged in user is "runner ".

expr

expression to run

log

a logical to indicate whether this message should be printed: Running scheduled task at...

ref_time

time to use for reference, defaults to Sys.time()

Details

The Windows Task Scheduler must be set up to use this function. Most convenient is to:

Create an R file such as R_cron.R with calls to schedule_task()

Create a batch file such as R_cron.bat that runs R_cron.R with R CMD BATCH

Set up a Task Scheduler task that runs R_cron.bat every minute

Examples

count_it <- function() {
  1 + 1
}

# units:      M  H  d  m  wd
schedule_task(., ., ., ., ., "user", count_it()) # every minute
schedule_task(0, ., ., ., ., "user", count_it()) # start of each hour
schedule_task(0, 7, ., ., ., "user", count_it()) # everyday at 7h00
schedule_task(0, 7, 1, ., ., "user", count_it()) # first day of month at 7h00
schedule_task(0, 7, ., 1, ., "user", count_it()) # everyday day in January at 7h00
schedule_task(0, 7, ., ., 1, "user", count_it()) # every Monday at 7h00

# every second Monday of the month at 7h00
# (= alleen maandag tussen 8-14e dag in de maand)
schedule_task(0, 7, c(8:14), ., 1, "user", count_it())

# every year on 15th of April at 8h30 and 16h30
schedule_task(30, c(8, 16), 15, 4, ., "user", count_it())

# once per quarter at 8 on the first day of the month
schedule_task(0, 8, 1, c(1, 4, 7, 10), ., "user", count_it())