Connect to Microsoft Teams via Microsoft 365

These functions create a connection to Microsoft Teams via Microsoft 365 and saves the connection to the certeprojects package environment. The teams_*() functions allow to work with this connection.

The older teams() function relies on the httr package and requires Incoming Webhooks to be set in Teams.

teams_connect(
  team_name = read_secret("teams.name"),
  tenant = read_secret("mail.tenant"),
  error_on_fail = FALSE
)

teams_name(account = teams_connect())

teams_channels(account = teams_connect())

teams_view_sharepoint(channel, account = teams_connect())

teams_send_message(
  body,
  channel,
  content_type = c("text", "html"),
  attachments = NULL,
  account = teams_connect()
)

teams_upload(
  local_path,
  teams_path = basename(local_path),
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = teams_connect()
)

teams_download(
  teams_path,
  local_path = basename(teams_path),
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = teams_connect()
)

teams_import(
  teams_path,
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = teams_connect()
)

teams_open(teams_path, channel = NULL, account = teams_connect())

teams_get_link(
  teams_path,
  share_type = c("view", "edit"),
  expire_after = "1 month",
  password = NULL,
  channel = NULL,
  account = teams_connect()
)

teams(
  message = "",
  channel,
  title = NULL,
  subtitle = NULL,
  items = NULL,
  items_title = NULL,
  images = NULL,
  images_title = NULL,
  buttons = NULL,
  message_thumbnail = NULL,
  markdown = TRUE,
  webhooks_file = read_secret("teams.webhooks.file"),
  ...
)

teams_webhooks(webhooks_file = read_secret("teams.webhooks.file"))

Arguments

team_name

name of the team

tenant

the tenant to use for Microsoft365R::get_team()

error_on_fail

a logical to indicate whether an error must be thrown if no connection can be made

account

a Microsoft 365 account to use for looking up properties. This has to be an object as returned by teams_connect() or Microsoft365R::get_team().

channel

channel to send message to, run teams_webhooks() for list of valid channels

body

text of the message

content_type

type of content, must be "text" or "html"

attachments

vector of file locations of attachments to add to the message

local_path

file location on the local system, can also be a data.frame which will then be saved locally to the temp folder first

teams_path

file location in Microsoft Teams, may also contain the channel name if channel is NULL, e.g., teams_path = "channel name/test.xlsx"

card_number

Trello card number

share_type

type of share, must be "view" (default) or "edit"

expire_after

time span after which the share link expires, defaults to "1 month", can also be e.g. "7 days"

password

password to set for share link, defaults to blank

message

message to send, supports markdown. Can also be a data.frame, which will be transformed using certestyle::plain_html_table().

title

title for message

subtitle

subtitle for message

items

text items for in message, a named vector

items_title

title to show above items

images

URLs of images to add to card. Local images will be uploaded to Imgur.com, or will will be transformed using base64enc::base64encode() if that fails.

images_title

title to show above images

buttons

buttons to show on card. Like items, this should be a named vector where the values are URLs (starting with http:// or https://) or mail links (starting with mailto:)

message_thumbnail

image to show on card

markdown

turn markdown support on/off

webhooks_file

the file locations of a YAML file with MS Teams webhooks in the form channel: "URL"

...

not used at the moment

Examples

if (FALSE) {

teams_upload("myfile.docx", channel = "My Channel")
teams_upload("myfile.docx", "my channel/my folder/test.docx")

# also supports data frames, they will be saved locally in a temp folder
teams_upload(mtcars, channel = "My Channel")
mtcars |> 
  teams_upload(channel = "My Channel")
  
  
# integrates with department projects:

# upload "myfile.docx" from the project folder of p123 to "My Channel":
teams_upload("myfile.docx", channel = "My Channel", card_number = 123)
# download "myfile.docx" from "My Channel" to the project folder of p123:
teams_download("myfile.docx", channel = "My Channel", card_number = 123)
teams_download("mychannel/myfile.docx", card_number = 123)

# direct import from Teams requires the 'certetoolbox' package
x <- teams_import("mychannel/myfile.docx")
x <- teams_import("mychannel/myfile.docx", card_number = 123)
x <- teams_import("myfile.docx", channel = "My Channel", card_number = 123)

# open a file in Excel Online
teams_open("test.xlsx", "My Channel")
teams_open("my channel/test.xlsx") # shorter version, tries to find channel
}
if (FALSE) {

# send message using Webhooks
teams("test message", channel = "My Channel")
# send message using Microsoft 365 login
teams_send_message("test message", channel = "My Channel")
}