Work with Trello Cards

This functions all call the Trello api.

trello_upload(
  title,
  desc = "",
  requested_by = "",
  project_path = "",
  member = Sys.getenv("R_USERNAME"),
  prio = read_secret("trello.default.prio"),
  duedate = "",
  checklist = "",
  checklist_name = read_secret("trello.default.checklist"),
  comments = "",
  attachments = "",
  list = read_secret("trello.default.list"),
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token"),
  ...
)

trello_credentials(x = c("member", "key", "token"))

trello_getboards(
  username = trello_credentials("member"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getboardname(
  board = read_secret("trello.default.board"),
  username = trello_credentials("member"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getlists(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getmembers(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getcards(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getlabels(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getcustomfields(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getchecklists(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getcomments(
  card_id,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_searchcard(
  x,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_openboard(board = read_secret("trello.default.board"))

trello_opencard()

trello_get_card_property(
  card_number,
  property,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_card_id(
  card_number,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_setcomment(
  card_id,
  comment,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_setmembers(
  card_id,
  member,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_setdeadline(
  card_id,
  duedate,
  duecomplete = FALSE,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_addtask(
  card_id,
  new_items = NULL,
  checklist_name = "Taken",
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_settask_state(
  card_id,
  checkitem_id,
  new_value = TRUE,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_movecard(
  card_id,
  list_id,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

Arguments

title
card title

desc
card description

requested_by
name of requester

project_path
path to project to which this card belongs to

member
Trello member, full name

prio
priority, must be label name

duedate
a Date object

checklist
items for checklist, can be character vector

checklist_name
name of the checklist

comments
comment to put on card, can be character vector

attachments
attachments to put on card

list
Trello list

board
Trello board

key
Trello API key

token
Trello API token

...
arguments passed on to methods

x
search string

username
Trello username

card_id
Trello card ID (not the card number)

card_number
Trello card number

property
property to take from card

comment
the comment to set, supports markdown

duecomplete
a logical to indicate whether due date is already completed

new_items
character vector of checklist items

checkitem_id
checkitem ID of item in checklist

new_value
the new value to set: a logical to indicate the 'complete' state of the checklist item

list_id
ID of the Trello board list