Connect to Microsoft Teams via Microsoft 365

These functions use the connection to Microsoft Teams set up with connect_teams().

teams_projects_channel(
  projects_channel_id = read_secret("teams.projects.channel_id"),
  overwrite = FALSE,
  account = connect_teams()
)

teams_new_project(
  task,
  channel = teams_projects_channel(),
  planner = connect_planner()
)

teams_browse_project(
  task,
  channel = teams_projects_channel(),
  planner = connect_planner()
)

teams_download_project_file(
  file,
  task,
  channel = teams_projects_channel(),
  planner = connect_planner()
)

teams_get_project_file(
  file,
  task,
  channel = teams_projects_channel(),
  planner = connect_planner()
)

teams_open_project_analysis_file(
  task,
  channel = teams_projects_channel(),
  planner = connect_planner()
)

teams_render_project_file(
  file,
  task,
  output_file = NULL,
  fun = rmarkdown::render,
  ...,
  channel = teams_projects_channel(),
  planner = connect_planner()
)

teams_view_project_file(
  file,
  task,
  channel = teams_projects_channel(),
  planner = connect_planner()
)

teams_upload_project_file(
  files,
  task,
  channel = teams_projects_channel(),
  planner = connect_planner()
)

teams_name(account = connect_teams())

teams_channels_list(account = connect_teams(), plain = TRUE)

teams_view_sharepoint(channel, account = connect_teams())

teams_send_message(
  body,
  channel,
  content_type = c("text", "html"),
  attachments = NULL,
  account = connect_teams()
)

teams_upload(
  local_path,
  teams_path = basename(local_path),
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = connect_teams()
)

teams_download(
  teams_path,
  local_path = basename(teams_path),
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = connect_teams()
)

teams_import(
  teams_path,
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = connect_teams()
)

teams_open(teams_path, channel = NULL, account = connect_teams())

teams_get_link(
  teams_path,
  share_type = c("view", "edit"),
  expire_after = "1 month",
  password = NULL,
  channel = NULL,
  account = connect_teams()
)

Arguments

projects_channel_id

Teams channel ID of the projects

overwrite

a logical to overwrite an existing connection, useful for switching accounts

account

a Microsoft 365 account to use for looking up properties. This has to be an object as returned by connect_teams() or Microsoft365R::get_team().

task

any task title, task ID, or ms_plan_task object (e.g. from planner_task_find())

channel

a Teams folder object. This has to be an object as returned by teams_projects_channel().

planner

a Microsoft 365 account for Planner. This has to be an object as returned by connect_planner().

file

the file name to open

output_file

path of the output file

fun

function to use for rendering. Can be e.g. rmarkdown::render or quarto::quarto_render.

...

arguments passed on to fun

files

the files to upload

plain

return as plain names, not as Azure objects

body

text of the message

content_type

type of content, must be "text" or "html"

attachments

vector of file locations of attachments to add to the message

local_path

file location on the local system, can also be a data.frame which will then be saved locally to the temp folder first

teams_path

file location in Microsoft Teams, may also contain the channel name if channel is NULL, e.g., teams_path = "channel name/test.xlsx"

card_number

Planner card number

share_type

type of share, must be "view" (default) or "edit"

expire_after

time span after which the share link expires, defaults to "1 month", can also be e.g. "7 days"

password

password to set for share link, defaults to blank

Details

The teams_new_project() function:

Checks if there is a Planner task with the correct task title

Creates a new folder in Teams in the projects channel

Updates the task to contain the project folder URL as an attachment

The teams_download_project_file() function will download the given project file to a temporary location, and will return the path of this location. This makes it possible to use source(), rmarkdown::render() or quarto::quarto_render() using the teams_download_project_file() function as input.

The teams_render_project_file() function allows to render a Teams file. It downloads the Teams file using teams_download_project_file(), runs the rendering function set in fun, and uploads the resulting output file back to Teams using the same file name a file, but with the new file extension (such as pdf, html, or docx). It invisibly returns the temporary local file location, so that the output of teams_render_project_file() can be given to e.g. certemail::mail() as an attachment.

teams_send_message() can also take a data.frame, which will be converted to HTML with plain_html_table(). If the input is a vector length > 1, the input will be collapsed with linebreaks.

Examples

if (FALSE) {
# PROJECT-RELATED --------------------------------------------------------

# Project-related Teams function rely on existing Planner tasks.

# create a new project, which will be a folder in the Teams channel
# for this, the task 'My Planner task' must already exist
teams_new_project("My Planner task")

# the task 'My Planner task' will now contain the URL to the project

# upload a file there
teams_upload_project_file("analysis.Rmd", "My Planner task")

# render R markdown or Quarto from and to the cloud
teams_render_project_file("analysis.Rmd", "My Planner task")
# this will put the output file in the same Teams folder as 'analysis.Rmd'

# PROJECT-UNRELATED ------------------------------------------------------

teams_upload("myfile.docx", channel = "My Channel")
teams_upload("myfile.docx", "my channel/my folder/test.docx")

# also supports data frames, they will be saved locally in a temp folder
teams_upload(mtcars, channel = "My Channel")
mtcars |> 
  teams_upload(channel = "My Channel")
  
  
# integrates with department projects:

# upload "myfile.docx" from the project folder of p123 to "My Channel":
teams_upload("myfile.docx", channel = "My Channel", card_number = 123)
# download "myfile.docx" from "My Channel" to the project folder of p123:
teams_download("myfile.docx", channel = "My Channel", card_number = 123)
teams_download("mychannel/myfile.docx", card_number = 123)

# direct import from Teams requires the 'certetoolbox' package
x <- teams_import("mychannel/myfile.docx")
x <- teams_import("mychannel/myfile.docx", card_number = 123)
x <- teams_import("myfile.docx", channel = "My Channel", card_number = 123)

# open a file in Excel Online
teams_open("test.xlsx", "My Channel")
teams_open("my channel/test.xlsx") # shorter version, tries to find channel
}