Work with Trello Cards

These functions call the Trello API and are internally used by project_get_current_id() (and its RStudio addin) and the Shiny apps project_add() and project_edit().

trello_upload(
  title,
  desc = "",
  requested_by = "",
  project_path = "",
  member = trello_credentials("membername"),
  prio = read_secret("trello.default.prio"),
  duedate = "",
  checklist = "",
  checklist_name = read_secret("trello.default.checklist"),
  comments = "",
  attachments = "",
  list = read_secret("trello.default.list"),
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token"),
  ...
)

trello_credentials(x = c("member", "key", "token", "membername"))

trello_get_boards(
  username = trello_credentials("member"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_board_name(
  board = read_secret("trello.default.board"),
  username = trello_credentials("member"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_lists(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_members(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_cards(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_labels(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_customfields(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_checklists(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_comments(
  card_id,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_my_cards(
  board = read_secret("trello.default.board"),
  username = trello_credentials("member"),
  key = trello_credentials("key"),
  token = trello_credentials("token"),
  list = read_secret("trello.current.list"),
  to_console = TRUE
)

trello_search_any(
  x,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_search_card(
  x = NULL,
  return_all = FALSE,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_open_board(board = read_secret("trello.default.board"))

trello_open_card()

trello_get_card_property(
  card_number,
  property,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_card_id(
  card_number,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_set_comment(
  card_id,
  comment,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_set_members(
  card_id,
  member,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_set_deadline(
  card_id,
  duedate,
  duecomplete = FALSE,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_add_task(
  card_id,
  new_items = NULL,
  checklist_name = "Taken",
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_set_task_state(
  card_id,
  checkitem_id,
  new_value = TRUE,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_move_card(
  card_id,
  list_id,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

Arguments

title

card title

desc

card description

requested_by

name of requester

project_path

path to project to which this card belongs to

member

Trello member, full name

prio

priority, must be label name

duedate

a Date object

checklist

items for checklist, can be character vector

checklist_name

name of the checklist

comments

comment to put on card, can be character vector

attachments

attachments to put on card

list

Trello list

board

Trello board

key

Trello API key

token

Trello API token

...

arguments passed on to methods

x

search string

username

Trello username

card_id

Trello card ID (not the card number)

to_console

a logical to indicate whether the card data should be printed to the Console

return_all

a logical to indicate whether a named vector of short URLs must be returned (internally used by project_add()) instead of a double vector

card_number

Trello card number

property

property to take from card

comment

the comment to set, supports markdown

duecomplete

a logical to indicate whether due date is already completed

new_items

character vector of checklist items

checkitem_id

checkitem ID of item in checklist

new_value

the new value to set: a logical to indicate the 'complete' state of the checklist item

list_id

ID of the Trello board list

Details

The trello_upload() returns the newly created Trello card number if the upload was successful.

The function trello_search_any() returns all cards data as a data.frame and is used internally by trello_search_card().